Fixes BUG 117810 : URI::Title 1.901 is not indexed.
The problem is Dist::Zilla::Plugin::PkgVersion which does not insert an
our declarator in front of the version. This gives a compile time error
when using strict and at the same time also using the $VERSION in the module.
If we try to manually declare our $VERSION in the module,
Dist::Zilla::Plugin::PkgVersion will skip the module, and the module
will not be indexed on cpan. To solve, this we can use
Dist::Zilla::Plugin::OurPkgVersion